+++ fiddles/nativeScript/fiddle-0010-Puzzle/puzzle/src/client/app/frameworks/core/testing/index.ts
@@ -1,0 +1,8 @@
+export * from './mocks/window.mock';
+
+// providers
+export * from './providers/core';
+export * from './providers/http';
+export * from './providers/router';
+
+