+++ fiddles/nativeScript/fiddle-0010-Puzzle/puzzle/tools/env/env-config.interface.ts
@@ -1,0 +1,2 @@
+export { EnvConfig } from '../../src/client/app/frameworks/core/utils/config';
+