+++ fiddles/nativeScript/fiddle-0010-Puzzle/puzzle/tools/tasks/project/desktop.libs.ts
@@ -1,0 +1,24 @@
+import * as gulp from 'gulp';
+import { relative, join } from 'path';
+import Config from '../../config';
+var newer = require('gulp-newer');
+
+export = () => {
+  let src = [
+    'node_modules/@angular/**/*',
+    'node_modules/rxjs/**/*',
+    'node_modules/angulartics2/**/*',
+    'node_modules/lodash/**/*',
+    'node_modules/ng2-translate/**/*',
+    'node_modules/@ngrx/**/*'
+  ];
+
+  src.push(...Config.NPM_DEPENDENCIES.map(x => relative(Config.PROJECT_ROOT, x.src)));
+
+  return gulp.src(src, { base: 'node_modules' })
+    .pipe(newer({
+      dest: join(Config.APP_DEST + '/node_modules'),
+      map: function(path: String) { return path.replace('.ts', '.js').replace('.scss', '.css'); }
+    }))
+    .pipe(gulp.dest(join(Config.APP_DEST + '/node_modules')));
+};